Lab3 nf 2

+++ studentNumberContract/contracts/StudentContract.sol
@@ -1,13 +1,14 @@
+// SPDX-License-Identifier: MIT 
 pragma solidity >=0.5.0;
 contract StudentContract {
-    uint public studentNumber ;
+    uint256 public studentNumber ;
     address public student ;
-    mapping ( address => uint ) studentToStudentNumber ;
+    mapping ( address => uint256 ) studentToStudentNumber ;
     constructor () public {
         student = msg . sender ;
         studentToStudentNumber [ student ] = 0;
     }
-    function setStudentNumber ( uint _studentNumber ) public payable {
+    function setStudentNumber ( uint256 _studentNumber ) public payable {
 
         require(msg.value == 5400000000000000, "Vous devez specifier 5400000000000000 wei");
 
@@ -15,8 +16,7 @@ contract StudentContract {
         studentToStudentNumber [ student ] = _studentNumber ;
     }
 
-    function getStudentNumber ( ) public view returns (uint) {
+    function getStudentNumber ( ) public view returns (uint256) {
         return studentNumber;
     }
-
 }
